Refactor namespaces and improve authentication handling
Updated namespace from `EV.Application.Common.Interface` to `EV.Application.Common.Interfaces` across multiple files for consistency. Adjusted import statements to reflect the new structure.

Modified authentication token handling in `axios-http.service.ts`, including endpoint path changes and navigation event dispatching for unauthorized access. These changes enhance code clarity and improve the client-side authentication flow.
