+++ EV.Application/Common/Interfaces/IAuthService.cs
@@ -1,6 +1,6 @@
 ﻿using EV.Application.Identity.Commands;
 
-namespace EV.Application.Common.Interface
+namespace EV.Application.Common.Interfaces
 {
     public interface IAuthService
     {
